LayoutTests/fast/events/simulated-click-coords-expected.txt

CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
Ensure that simulated click on label elements uses the original mouse coordinates.

 Click Me
 Click Me
PASS: event.clientX is same for both real and simulated event.
PASS: event.clientY is same for both real and simulated event.
PASS: event.layerX is same for both real and simulated event.
PASS: event.layerY is same for both real and simulated event.
PASS: event.pageX is same for both real and simulated event.
PASS: event.pageY is same for both real and simulated event.
PASS: event.screenX is same for both real and simulated event.
PASS: event.screenY is same for both real and simulated event.
PASS: event.x is same for both real and simulated event.
PASS: event.y is same for both real and simulated event.

PASS: event.clientX is same for both real and simulated event.
PASS: event.clientY is same for both real and simulated event.
PASS: event.layerX is same for both real and simulated event.
PASS: event.layerY is same for both real and simulated event.
PASS: event.pageX is same for both real and simulated event.
PASS: event.pageY is same for both real and simulated event.
PASS: event.screenX is same for both real and simulated event.
PASS: event.screenY is same for both real and simulated event.
PASS: event.x is same for both real and simulated event.
PASS: event.y is same for both real and simulated event.
